+++ apps/hey-matter/test/package.test.ts
@@ -1,9 +1,5 @@
-import backend from "@hey-matter/backend/package.json" with {
-  type: "json",
-};
-import common from "@hey-matter/common/package.json" with {
-  type: "json",
-};
+import backend from "@hey-matter/backend/package.json" with { type: "json" };
+import common from "@hey-matter/common/package.json" with { type: "json" };
 import { mapValues, pickBy } from "lodash-es";
 import { describe, expect, it } from "vitest";
 import own from "../package.json" with { type: "json" };