+++ packages/common/src/index.ts
@@ -2,6 +2,7 @@ export * from "./bridge-data.js";
 export * from "./clusters/index.js";
 export * from "./domains/index.js";
 export * from "./endpoint-data.js";
+export * from "./entity-option.js";
 export * from "./home-assistant-device-registry.js";
 export * from "./home-assistant-domain.js";
 export * from "./home-assistant-entity-information.js";